refactor(channels/discord): P4 — extract interaction.rs (callbacks + pending store)
Move the interaction followup-credential store (PendingInteraction,
INTERACTION_TOKEN_TTL) and the REST callbacks (defer, reject, edit @original)
plus interaction_string_option into discord/interaction.rs. interaction_gate +
InteractionDenial stay in mod.rs (the gate is coupled to channel_passes_filter).
Behaviour-neutral: 161 discord tests pass.
